chore (tests): split storybook tests in batches
