+++ cypress/e2e/stories-3.cy.js
@@ -1,0 +1,8 @@
+import "cypress-network-idle";
+import { getFilteredStories, runBatch } from "../support/storyUtils";
+
+describe("Stories test - batch 3", () => {
+  const all = getFilteredStories();
+  const chunkSize = Math.ceil(all.length / 3);
+  runBatch(all.slice(chunkSize * 2));
+});
