--- cypress/e2e/stories-map.cy.js
@@ -1,10 +1,0 @@
-import stories from "../../docs/index.json";
-
-describe("Map Stories", () => {
-  Cypress.env(
-    "SPEC_FILTER",
-    "map,layercontrol,drawtools,geosearch,timecontrol",
-  );
-});
-
-import "./storybook-engine.js";
