--- cypress/e2e/stories-other.cy.js
@@ -1,8 +1,0 @@
-describe("Other Stories", () => {
-  Cypress.env(
-    "SPEC_FILTER",
-    "chart,feedback,itemfilter,jsonform,layout,stacinfo,storytelling",
-  );
-});
-
-import "./storybook-engine.js";
